reset and re-populate list of imports into symbol table each time a file parser is run
Qids of imported packages are per-file scoped, not per packages.
So they have to re-populated each time a file symbols are (re)processed.
For that reason, postponed data types of all files are re-processed first.
Then variables, then functions. In the worts cast the imported packages
are re-populated (3 x |files|).