fix *expr to reference instead of dereference
don't parse function body if it does not exist (e.g. function declaration only)

store function definition only if it does not exists (e.g. if a function is reprocessed it can be already stored in the symbol table)

set the GOROOT as the first path to search for go package; be more verbose