retrieve fields from any data type
E.g. the following definition is possible:
```go
type A []string

func (a*A) len() int {
	return len(*a)
}

type B struct {
	A
}

func main() {
	a := A{"a", "b"}
	b := B{a}
	fmt.Printf("len: %v", b.len())
}
```

The A is a list of strings with one method. So its field is the method.
The B embeds A so it has the same field as A has.