Store params and returns once all of them are processed
In case a function is in the following form

func(net, addr string) (net.Conn, error)

the net.Conn is incorrectly processed as the net id is thought to be func's first parameter of type string.
Thus, all the params and returns ids must be stored into the symbol table once all params and returns
are fully processed.